Changed directory logic, implemented routing and admin-only profile lists
