--- src/helpers/script.js
@@ -1,24 +1,0 @@
-$( document ).ready(function() {
-
-    // Hero Section FUll Height
-    $(function(){
-    	var windowH = $(window).height();
-    	var wrapperH = $('.hero').height();
-    	if(windowH > wrapperH) {                            
-    		$('.hero').css({'height':($(window).height())+'px'});
-    	}                                                                               
-    	$(window).resize(function(){
-    		var windowH = $(window).height();
-    		var wrapperH = $('.hero').height();
-    		var differenceH = windowH - wrapperH;
-    		var newH = wrapperH + differenceH;
-    		var truecontentH = $('#truecontent').height();
-    		if(windowH > truecontentH) {
-    			$('.hero').css('height', (newH)+'px');
-    		}
-
-    	})          
-    });
-
-
-});
